+++ src/aks-preview/azext_aks_preview/_params.py
@@ -123,6 +123,7 @@ def load_arguments(self, _):
         c.argument('appgw_subnet_id', options_list=['--appgw-subnet-id'], arg_group='Application Gateway')
         c.argument('appgw_watch_namespace', options_list=['--appgw-watch-namespace'], arg_group='Application Gateway')
         c.argument('aci_subnet_name', type=str)
+        c.argument('enable_encryption_at_host', options_list=['--enable-encryption-at-host'], action='store_true')
         c.argument('yes', options_list=['--yes', '-y'], help='Do not prompt for confirmation.', action='store_true')
 
     with self.argument_context('aks update') as c:
@@ -187,6 +188,7 @@ def load_arguments(self, _):
             c.argument('pod_subnet_id', type=str, validator=validate_pod_subnet_id)
             c.argument('kubelet_config', type=str)
             c.argument('linux_os_config', type=str)
+            c.argument('enable_encryption_at_host', options_list=['--enable-encryption-at-host'], action='store_true')
 
     for scope in ['aks nodepool show', 'aks nodepool delete', 'aks nodepool scale', 'aks nodepool upgrade', 'aks nodepool update']:
         with self.argument_context(scope) as c: