+++ src/aks-preview/HISTORY.md
@@ -5,6 +5,7 @@ Release History
 0.4.73
 +++++
 * Update to use 2020-12-01 api-version
+* Add argument '--enable-encryption-at-host'
 
 0.4.72
 ++++++